Updating SolidParticleSystem class to v4.1.0-beta.13

+++ src/BabylonCpp/include/babylon/core/structs.h
@@ -208,7 +208,7 @@ struct NodeCache {
 }; // end of struct NodeCache
 
 struct PickedParticle {
-  unsigned int idx;
+  size_t idx;
   size_t faceId;
 }; // end of struct PickedParticle
