+++ src/BabylonCpp/include/babylon/particles/model_shape.h
@@ -26,13 +26,12 @@ public:
    * for a model to replicate particles from by the SPS. SPS internal tool, don't use it manually.
    * @hidden
    */
-  ModelShape(
-    int id, const std::vector<Vector3>& shape, const IndicesArray& indices,
-    const Float32Array& normals, const Float32Array& colors, const Float32Array& shapeUV,
-    const std::function<void(SolidParticle* particle, unsigned int i, unsigned int s)>& posFunction,
-    const std::function<void(SolidParticle* particle, const Vector3& vertex, unsigned int i)>&
-      vtxFunction,
-    const MaterialPtr& material);
+  ModelShape(int id, const std::vector<Vector3>& shape, const IndicesArray& indices,
+             const Float32Array& normals, const Float32Array& colors, const Float32Array& shapeUV,
+             const std::function<void(SolidParticle* particle, size_t i, size_t s)>& posFunction,
+             const std::function<void(SolidParticle* particle, const Vector3& vertex,
+                                      unsigned int i)>& vtxFunction,
+             const MaterialPtr& material);
   ~ModelShape(); // = default
 
 public:
@@ -75,7 +74,7 @@ public:
    * Custom position function (internal use)
    * Hidden
    */
-  std::function<void(SolidParticle* particle, unsigned int i, unsigned int s)> _positionFunction;
+  std::function<void(SolidParticle* particle, size_t i, size_t s)> _positionFunction;
   /**
    * Custom vertex function (internal use)
    * Hidden
