+++ src/BabylonCpp/src/particles/model_shape.cpp
@@ -7,7 +7,7 @@ namespace BABYLON {
 ModelShape::ModelShape(
   int id, const std::vector<Vector3>& shape, const IndicesArray& indices,
   const Float32Array& normals, const Float32Array& colors, const Float32Array& shapeUV,
-  const std::function<void(SolidParticle* particle, unsigned int i, unsigned int s)>& posFunction,
+  const std::function<void(SolidParticle* particle, size_t i, size_t s)>& posFunction,
   const std::function<void(SolidParticle* particle, const Vector3& vertex, unsigned int i)>&
     vtxFunction,
   const MaterialPtr& material)
